+++ src/main/java/nextstep/subway/path/domain/graph/DijkstraPath.java
@@ -1,4 +1,4 @@
-package nextstep.subway.path.domain;
+package nextstep.subway.path.domain.graph;
 
 import nextstep.subway.station.domain.Station;
 import org.jgrapht.Graph;
