+++ src/main/java/nextstep/subway/path/domain/graph/StationGraph.java
@@ -1,4 +1,4 @@
-package nextstep.subway.path.domain;
+package nextstep.subway.path.domain.graph;
 
 import nextstep.subway.line.domain.Line;
 import nextstep.subway.line.domain.Section;
