+++ src/test/java/nextstep/subway/path/domain/PathFinderTest.java
@@ -3,6 +3,9 @@ package nextstep.subway.path.domain;
 import nextstep.subway.line.domain.Distance;
 import nextstep.subway.line.domain.Line;
 import nextstep.subway.line.domain.Section;
+import nextstep.subway.path.domain.graph.DijkstraPath;
+import nextstep.subway.path.domain.graph.PathAlgorithm;
+import nextstep.subway.path.domain.graph.StationGraph;
 import nextstep.subway.path.exception.InvalidFindShortestPathException;
 import nextstep.subway.station.domain.Station;
 import org.junit.jupiter.api.BeforeEach;
